app should start with just REST if graphql and notifications off
